Movement Complete
Slight bugs with sprites when changing direction.  Main multidirectional movement completed.

+++ HappySumo/SnakeGame/GameObject.cpp
@@ -1,6 +1,6 @@
 #include "GameObject.h"
 
-GameObject::GameObject(sf::Vector2f new_position, sf::Color new_color, int new_Size) : currentPosition (new_position), color (new_color), size (new_Size)
+GameObject::GameObject(sf::Vector2f new_position, sf::Color new_color, float new_Size) : currentPosition (new_position), color (new_color), size (new_Size)
 {
 }
